moved unit tests to correct tests folder
