+++ src/main/java/module-info.java
@@ -14,11 +14,8 @@ module truck.truckmanagement {
     exports truck.truckmanagement;
     opens truck.truckmanagement.Controller;
     exports truck.truckmanagement.Controller;
-    //opens truck.truckmanagement.Controller to javafx.fxml;
-    //exports truck.truckmanagement.Controller to javafx.fxml;
 
     opens truck.truckmanagement.Model to org.hibernate.orm.core;
 
     opens truck.truckmanagement.TableParameters to javafx.base;
-    exports truck.truckmanagement.Tests;
 }
